@@ -1,49 +1,46 @@
-
 export const cornerOptions = [
-    "corner of the internet",
-    "part of the web",
-    "slice of the digital world",
-    "spot in cyberspace",
-    "digital realm",
-    "virtual space",
-    "web haven",
-    "section of the net",
-    "patch of cyberspace",
-  ];
-
+  "corner of the internet",
+  "part of the web",
+  "slice of the digital world",
+  "spot in cyberspace",
+  "digital realm",
+  "virtual space",
+  "web haven",
+  "section of the net",
+  "patch of cyberspace",
+];
 
 export const greetings = [
-    { text: "Hey", flag: null },  
-    { text: "Hola", flag: "🇪🇸" }, 
-    { text: "Bonjour", flag: "🇫🇷" },
-    { text: "Merhaba", flag: "🇹🇷" },
-    { text: "Hallo", flag: "🇩🇪" }, 
-    { text: "Ciao", flag: "🇮🇹" }, 
-    { text: "Olá", flag: "🇧🇷" }, 
-    { text: "नमस्ते", flag: "🇮🇳" }, 
-    { text: "سلام", flag: "🇵🇰" },
-    { text: "こんにちは", flag: "🇯🇵" }, 
-    { text: "안녕하세요", flag: "🇰🇷" },
-
-  ];
+  { text: "Hey", flag: null },
+  { text: "Hola", flag: "🇪🇸" },
+  { text: "Bonjour", flag: "🇫🇷" },
+  { text: "Merhaba", flag: "🇹🇷" },
+  { text: "Hallo", flag: "🇩🇪" },
+  { text: "Ciao", flag: "🇮🇹" },
+  { text: "Olá", flag: "🇧🇷" },
+  { text: "नमस्ते", flag: "🇮🇳" },
+  { text: "سلام", flag: "🇵🇰" },
+  { text: "こんにちは", flag: "🇯🇵" },
+  { text: "안녕하세요", flag: "🇰🇷" },
+];
 
 export const hobbies = [
-    "jamming on the guitar",
-    "figuring out how not to fall off a snowboard",
-    "totally not coding anyway - I swear",
-    "trying not to get lost while camping",
-    "befriending a stray cat",
-    "getting stuck halfway up a climbing wall but loving it anyway",
-    "attempting to cook something edible",
-    "dreaming of coding from/for space someday",
-    "definitely not getting lost hiking",
-    "attempting to hit Grand Champ in Rocket League",
-    "creating another Spotify playlist",
-    "learning a new hobby just to forget it a week later",
-    "staring at the stars through a telescope",
-    "finding new ways to chase the stars - literally and metaphorically",
-    "totally not procrastinating by adding to this list"
-]
+  "jamming on the guitar",
+  "figuring out how not to fall off a snowboard",
+  "totally not coding anyway - I swear",
+  "trying not to get lost while camping",
+  "befriending a stray cat",
+  "getting stuck halfway up a climbing wall but loving it anyway",
+  "attempting to cook something edible",
+  "dreaming of coding from/for space someday",
+  "definitely not getting lost hiking",
+  "attempting to hit Grand Champ in Rocket League",
+  "creating another Spotify playlist",
+  "learning a new hobby just to forget it a week later",
+  "staring at the stars through a telescope",
+  "finding new ways to chase the stars - literally and metaphorically",
+  "totally not procrastinating by adding to this list",
+];
 
 export const shortGoals = [
   "learning how to play the piano 🎹",
@@ -53,7 +50,7 @@ export const shortGoals = [
   "cooking a new dish every week 😋",
   "go scuba diving 🤿",
   "mastering snowboarding 🏂",
-]
+];
 
 export const longGoals = [
   "start my own company 🧑‍💻",
@@ -62,7 +59,7 @@ export const longGoals = [
   "help advance the human civilization 📡",
   "eradicate world hunger 💪",
   "learn how to fly a plane ✈️",
-]
+];
 
 export const currently = [
   "Currently",
@@ -70,7 +67,7 @@ export const currently = [
   "At the moment",
   "As of now",
   "Lately",
-]
+];
 
 export const previously = [
   "Previously",
@@ -78,7 +75,7 @@ export const previously = [
   "In the past",
   "Earlier",
   "Formerly",
-]
+];
 
 export const achievements = [
   "Achievements",
@@ -86,4 +83,4 @@ export const achievements = [
   "Awards",
   "Successes",
   "Triumphs",
-]
+];
