@@ -1,18 +1,21 @@
-import React from 'react'
+import React from "react";
 
-const ChangeWord = ({list}) => {
-    const [currentWord, setCurrentWord] = React.useState(list[0]);
+const ChangeWord = ({ list }) => {
+  const [currentWord, setCurrentWord] = React.useState(list[0]);
 
-    const handleWordClick = () => {
-        const nextIndex = (list.indexOf(currentWord) + 1) % list.length;
-        setCurrentWord(list[nextIndex]);
-    }
+  const handleWordClick = () => {
+    const nextIndex = (list.indexOf(currentWord) + 1) % list.length;
+    setCurrentWord(list[nextIndex]);
+  };
 
   return (
-    <span className='underline cursor-pointer font-semibold select-none' onClick={handleWordClick}>
-        {currentWord}
+    <span
+      className="underline cursor-pointer font-semibold select-none"
+      onClick={handleWordClick}
+    >
+      {currentWord}
     </span>
-  )
-}
+  );
+};
 
-export default ChangeWord
+export default ChangeWord;
