@@ -15,9 +15,9 @@ const Layout = () => {
         <main className="w-full flex-1 flex md:max-w-[43%] justify-items-center mobile:max-w-[80%] mx-auto mb-4">
           <Outlet />
         </main>
-        
+
         {/* Footer */}
-        <div className="">
+        <div className="mt-2">
           <Footer />
         </div>
       </div>
