@@ -1,21 +1,21 @@
-import React, { useState, useEffect } from 'react';
+import { useState, useEffect } from "react";
 
 const Quotes = () => {
-  const [quoteData, setQuoteData] = useState({ quote: '', author: '' });
+  const [quoteData, setQuoteData] = useState({ quote: "", author: "" });
   const [isLoading, setIsLoading] = useState(true);
 
   const fetchRandomQuote = async () => {
     setIsLoading(true);
     try {
-      const response = await fetch('/.netlify/functions/quote');
+      const response = await fetch("/.netlify/functions/quote");
       const data = await response.json();
       setQuoteData({
-        quote: data.quote || 'No quote available',
-        author: data.author || 'Unknown',
+        quote: data.quote || "No quote available",
+        author: data.author || "Unknown",
       });
     } catch (error) {
-      console.error('Error fetching the quote:', error);
-      setQuoteData({ quote: 'Error fetching quote', author: 'Unknown' });
+      console.error("Error fetching the quote:", error);
+      setQuoteData({ quote: "Error fetching quote", author: "Unknown" });
     } finally {
       setIsLoading(false);
     }
@@ -32,7 +32,9 @@ const Quotes = () => {
       ) : (
         <>
           <div className="quote-text">
-            <p className="italic text-lg mb-4 font-semibold">"{quoteData.quote}"</p>
+            <p className="italic text-lg mb-4 font-semibold">
+              "{quoteData.quote}"
+            </p>
           </div>
           <div className="quote-author">
             <p className="text-xl font-light">- {quoteData.author}</p>
