@@ -1,17 +1,16 @@
-import React, { useState, useEffect } from 'react'
-import { useTheme } from '../context/ThemeContext';
+import { useTheme } from "../context/ThemeContext";
 
 const ThemeToggle = () => {
-    const { isDark, toggleTheme } = useTheme();
+  const { isDark, toggleTheme } = useTheme();
 
   return (
     <button
-        onClick={toggleTheme}
-        className="text-textlight dark:text-textdark transition-all duration-300 ease-in-out"
+      onClick={toggleTheme}
+      className="text-textlight dark:text-textdark transition-all duration-300 ease-in-out"
     >
-        {isDark ? 'moon' : 'sun'}
+      {isDark ? "moon" : "sun"}
     </button>
-  )
-}
+  );
+};
 
 export default ThemeToggle;
