@@ -1,4 +1,4 @@
-const nodemailer = require('nodemailer');
+const nodemailer = require("nodemailer");
 
 exports.handler = async (event, context) => {
   // Parse request body
@@ -12,7 +12,7 @@ exports.handler = async (event, context) => {
   `;
 
   const transporter = nodemailer.createTransport({
-    service: 'gmail',
+    service: "gmail",
     auth: {
       user: process.env.EMAIL_USER, // Set EMAIL_USER in Netlify's environment variables
       pass: process.env.EMAIL_PASS, // Set EMAIL_PASS in Netlify's environment variables
@@ -30,13 +30,13 @@ exports.handler = async (event, context) => {
     await transporter.sendMail(mailOptions);
     return {
       statusCode: 200,
-      body: JSON.stringify({ message: 'Email sent' }),
+      body: JSON.stringify({ message: "Email sent" }),
     };
   } catch (error) {
-    console.error('Error sending email:', error);
+    console.error("Error sending email:", error);
     return {
       statusCode: 500,
-      body: JSON.stringify({ message: 'Error sending email' }),
+      body: JSON.stringify({ message: "Error sending email" }),
     };
   }
 };
